adding a helper functions folder
